requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
flask-restplus==0.12.1
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
psycopg2==2.7.6.1
py==1.5.4
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pytest==3.7.1
pytest-cov==2.6.0
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
